Word-clock-bot
A bot built on the Microsoft Bot Framework. This bot will look for
commands such as "time". It will tell you what time is it according to
the server time in English vocabulary.  There are some random twists to
it as well.
==============

- Bot will first ask for user name. It will remember till end of the
session.
- Bot will use regex to look for "time" command.
- Bot will respond in words, but sometimes it will give you some
personal responds to give some sort of character.
- To deactivate the bot just to say "bye" or "goodbye" to it.